remove interface SQLRewriteEngine and SQLRouteRewriteEngine
